Break out into modules and add tests
