+++ slack_to_discord/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+from slack_to_discord.importer import run_import
+from slack_to_discord.http_stream import SeekableHTTPStream
